Pizza-Agent/agent/requirements.txt
streamlit
python-dotenv
pydantic
openai